# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VentasFix",
        "description": "Microservicio de carro de compra de VentasFix: usuarios, productos, clientes y dashboard. Autenticación con JWT (Bearer).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "2.1 Listar todos los productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Listado obtenido"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "2.3 Agregar un nuevo producto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "name",
                                    "short_description",
                                    "long_description",
                                    "image",
                                    "net_price",
                                    "current_stock",
                                    "minimum_stock",
                                    "low_stock",
                                    "high_stock"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "AUD-050"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Audífonos inalámbricos"
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "example": "Cancelación de ruido"
                                    },
                                    "long_description": {
                                        "type": "string",
                                        "example": "Audífonos over-ear con 30 h de batería y estuche de carga."
                                    },
                                    "image": {
                                        "description": "Ruta o URL de la única imagen del producto",
                                        "type": "string",
                                        "example": "images/logo.png"
                                    },
                                    "net_price": {
                                        "description": "Precio neto en CLP. El precio de venta se calcula solo (+19% IVA)",
                                        "type": "integer",
                                        "example": 59990
                                    },
                                    "current_stock": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "minimum_stock": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "low_stock": {
                                        "description": "Debe ser >= minimum_stock",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "high_stock": {
                                        "description": "Debe ser >= low_stock",
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado (incluye sale_price calculado)"
                    },
                    "422": {
                        "description": "Datos inválidos o vacíos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "2.2 Obtener un producto por su ID",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado"
                    },
                    "404": {
                        "description": "No existe un producto con ese id"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "2.4 Actualizar un producto por su ID (PUT o PATCH)",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "TEC-001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Teclado mecánico TKL Pro"
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "example": "Switches red"
                                    },
                                    "long_description": {
                                        "type": "string",
                                        "example": "Descripción larga actualizada."
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "images/logo.png"
                                    },
                                    "net_price": {
                                        "description": "Si cambia, sale_price se recalcula",
                                        "type": "integer",
                                        "example": 49990
                                    },
                                    "current_stock": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "minimum_stock": {
                                        "description": "Los 3 umbrales van juntos",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "low_stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "high_stock": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "404": {
                        "description": "No existe un producto con ese id"
                    },
                    "422": {
                        "description": "Datos inválidos o vacíos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "2.5 Eliminar un producto por su ID",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "404": {
                        "description": "No existe un producto con ese id"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Pega el access_token q devuelve /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}